Fixes AppsIndexerManagerServiceTest
The main issue is that we confused uid with UserHandles, and set EXTRA_UID to the UserHandle.

The receiver in apps indexer would get extra value we passed in the Intent, then call getUserHandleForUid, which essentially divides the value by 100K (PER_USER_RANGE). It would then use this to determine which user to run apps indexer for. We didn't see this when we passed 0 as the UID, as it would get divided by 100K and map to 0 anyways. However, when we passed 10 (default userhandle on HSUM devices), that would also get mapped to 0. So part of this test would check apps indexer for user 10, and part of it would test it for user 0.

This isn't an issue outside of this test, as real "ACTION_PACKAGE_ADDED" and related events sent by the system have EXTRA_UID set correctly.

Test: AppsIndexerManagerServiceTest
Bug: 383091450
Flag: TEST_ONLY
Change-Id: I6c9e396ecc71cc3c22b07c9ddfbf4052f978f383