Add AppFunction tests to AppSearch presubmits.
AppFunction relies on some AppSearch APIs, but was not previously tested.
This cause some tests to fail when we made changes to AppSearch APIs,
causing failures on git_main.

This CL adds AppFunction tests to presubmits so that these tests get run
whenever we make changes to AppSearch, preventing failures.

Also sorted tests in alphabetical order.

Test: None. Adding test suite to TEST_MAPPING.
Change-Id: I24a6a00dc240e76aa4dbade57276a8e0b506c927