Changes AppsIndexerImpl to check for app update ts difference
Previously this was checking if the last update was strictly GREATER
than what was recorded in AppSearch. We will now check if the last
update timestamp is DIFFERENT from what was recorded in AppSearch. This
handles the case where the system time is modified, and the last update
timestamp somehow ends up lower then whats in AppSearch

Flag: EXEMPT bugfix
Bug: 275592563
Test: presubmit
Test: testAppsIndexerImpl_incrementalPut_differentTimestamp_reindexes
Change-Id: I16c65737e64ae61fdf269bdf3ec92b29511100d1