Refactor AppSearch RevocableFileDescriptor and store.
This change RevocableFileDescriptorStore to be an abstract class so it
could be reused in framework.
This change also split cached opened fd for read and write to apply
different limitation on them. We should allow user to open multiple
fd for read but only 1 fd for write each file.

Bug:382592895
Test: presubmit
Flag: com.android.appsearch.flags.enable_blob_store
Change-Id: I145453bf3030bb5ed883259741622a10cd7f3657