Fix testAppSearchRateLimit_rateLimitChangeToHigherLimit test flake
There's a race condition where the set schema task has not been removed yet due to the asynchronous nature of the rate limited executor before the search task is scheduled, causing the search task to be rate limited. I've changed the api cost of search so it won't get rate limited when this happens. This does not seem to change the semantics of the test.

Bug: 390832000
Test: AppSearchManagerServiceTest.java
Flag: TEST_ONLY
Change-Id: Id465415a11075242b93a8c672a68f33c31ffdb0d