AppSearch/IsolatedStorageService: Disable Backup/Restore
AppSearch currently does not support backup/restore. The pVM is
tied to a unique device secret, and thus if the pVM is restored on
a different device, the VM would be invalid. Therefore it does not
make sense to support backup/restore for the Isolated Storage
Service apk. Disable support for backup/restore for ISS.

Test: AppSearch CTS Tests
Bug: 384766593
Flag: com.android.appsearch.flags.enable_isolated_storage
Change-Id: I7e42b3486c303f047019ff2af0aeed65221d81c4