Eliminate blocking for calls to getInstanceOrNull.
The lock guarding the map of user instances is held during user creation. Callers of getInstanceOrNull are, by definition, tolerant of situations in which the instance is not available. Therefore, the better behavior is to simply return null in those cases rather than blocking.

For some further details, see b/407104133#comment6

Bug: 407104133
Flag: EXEMPT trivial bug fix
Test: presubmits
Change-Id: I95bf57c266bb736a0f59aff1035201343b586739