appsearch: icing combined
Only use one library for icing.

- libicing_anywhere works in system server or in VM
- libappsearchservice is still separate, but converted to C so only 51K.

Note: before doing this, the library inside and outside the APK are
confirmed to be the same:

$ cd $ANDROID_PRODUCT_OUT/apex/com.android.appsearch/
$ unzip priv-app/com.android.appsearch.apk@MAIN/com.android.appsearch.apk.apk
...
$ diff ./lib64/libicing_anywhere.so ./lib/arm64-v8a/libicing_anywhere.so && echo SAME!
SAME!

After this change, there is no longer a library in
com.android.appsearch.apk.apk. There is only a library
in the APEX.

Now that we are mounting the APEX in the VM, the DICE chain is
invalidated, so we delete the old VM instance, and our VM
instance is renamed.

Bug: 406856088
Test: boot w/ VM on, CTS
Test: boot w/o VM, CTS
Test: update to this build, crosvm doesn't crash
Test: reboot multiple times, no error from deleting old
    VM instance.

Change-Id: Ic1acf5250f5a75c4152b9804cbed748bf5ca5dee