Fix testScheduleAndRunJob_unsetIndexerType_indexerTypeIsSet flake
Most likely, the job rescheduling is failing, and according to the android docs, one cause of this is if an app schedules jobs too quickly. Looking at logs, it looks like the tests that failed had the initial jobs run within 20-30 ms of being scheduled. A passing test has the job run within 40-50 ms of being scheduled. To be safe, I've set the minimum latency much higher (200 ms). Unfortunately, there doesn't seem to be an obvious constant for the minimum job scheduling interval.

Bug: 416918440
Test: ContactsIndexerMaintenanceTest.java
Flag: TEST_ONLY
Change-Id: I6ff6737d62077260d6b7da7f81e87f6b1eaf6898