Populate IcuDataStatusCode with new FailureStage.
We've added a new field to the result returned from native
initialization. This field will indicate roughly where we've failed. In
this change we store that value in the pre-existing IcuDataStatusCode
field.

In a followup, we should add a proper atom field for this value but as
this is needed to diagnose dogfood issues ASAP, we're reusing a field.

Bug: 415387509
Flag: EXEMPT trivial log change (also only runs with pVM enabled)
Test: presubmits
Change-Id: I114102d8509abf75cc5579eafd2a1eaf71c2f945