AppSearch/IsolatedStorageService: increase vm idle timeout
use 20s to make the vm idle ~90% of the time

metrics and analysis from @tjbarron@google.com:

In prod, p50 is 30 calls per day and p99 is 300 calls per day. In dogfood, p50 is also about 30 and p99 is ~1500. So that would mean that, at p99 in dogfood, if API calls were evenly spaced out through out the day, then you'd have 1 API call every 57.6s. So with a timeout of 60s, then I guess in that particular case, the balloon would never run. Now it's almost certainly not the case that API calls are evenly distributed. Let's instead assume that all API calls come in batches of 5 each 2s apart. That would mean that we'd get the beginning of a batch every 288s. With a 60s timeout, we'd start the balloon with 220s before the next batch begins. So the VM would be idle ~76% of the time. With the 5s timeout, the VM would have been idle ~95% of the time. Perhaps we start by saying, let's try to estimate a timeout that would produce an estimate of 90% idle? Which I think would be ~20s? Doing that same calculation for the ~30 calls that the median user gets per day would mean the difference between 99.9% idle and 99.5% idle.

Test: appsearch cts
Bug: 415310309
Flag: com.android.appsearch.flags.enable_isolated_storage
Change-Id: I506752e3bb81c3ef341857b3f6e9aece01473537