Don't use Isolated Storage where it isn't appropriate
To meet user and enterprise requirements, when Android claims to
have deleted data, we need to ensure that the data is actually
inaccessible.

Android has two primary, overlapping mechanisms to ensure that
data is inaccessible when a profile is deleted:

* Deletion of the underlying files (eg rm -rf /data/system_ce/$USER_ID).
* Deletion of the encryption key ("CE key") associated with the profile.

Either one of these mechanisms alone is sufficient to make data
inaccessible. Both are applied for redundancy and to reclaim storage
space.

Prior to this change, the Isolated Storage service was storing data
associated with one user encrypted with the CE key of a different user.
As a consequence, deleting the CE key of a user may not make their data
inaccessible.

This change limits the scope of the Isolated Storage service to
only store data associated with the User who's CE key is encrypting
the isolated storage data. Other data will be stored in the
conventionally protected /data/system_ce/#/appsearch directory. The
security properties of Isolated Storage and conventional CE storage
are identical, with the added robustness of guaranteed deletion of
enterprise data.

Change-Id: I85072d6e46501c969962f945a272671832afc71d
Fixes: 416532681
Flag: EXEMPT bugfix
Test: atest AppSearchSessionCtsTest
Test: atest --rebuild-module-info CtsAppSearchTestCases AppSearchServicesTests AppSearchMockingServicesTests AppSearchCoreTests
Test: atest --rebuild-module-info CtsAppSearchHostTestCases