testIsolatedStorageNotAvailable: reuse user0
The AppSearchManagerServiceTest#testIsolatedStorageNotAvailable
test created a new UserHandle when testing that a mocked isolated
storage is not available. This is unnecessary. Instead, we can
make the changes which enable isolated storage and clear the
AppSearch instance cache, which largely has the same effect and
allows us to test with a properly configured user0.

Test: atest com.android.server.appsearch.AppSearchManagerServiceTest#testIsolatedStorageNotAvailable
Flag: EXEMPT test only change
Change-Id: Ibdba03b7646a287e79e9d53203f21a970c24937a