Re-create VM if actual config is not compatible with expected one
Recently we've added an option for VMs to specify which encrypted store
mode they want to use via specifying
`android.system.virtualmachine.ENCRYPTED_STORE_MODE` property in the
AndroidManifest.xml of the VM owner app. The value is set to
VirtualMachineConfig at the moment when it is built.

We are in process of making encrypted store mode part of the
VirtualMachineConfig#isCompatibleWith check, meaning that appsearch VM
needs to be deleted and re-created.

Bug: 423625345
Test: remove ENCRYPTED_STORE_MODE property from the manifest
Test: build && install appsearch APEX
Test: check logs contain line about re-creating the VM
Flag: EXEMPT, bug fix
Change-Id: I4116927724ddef35f07019284857a204b0db64a0