Add ART profile to AppSearch service.
This CL adds an ART profile to the AppSearch service. The profile was
generated by profiling the AppSearch service on Droidfood population.

The ART profile is expected to reduce the size of the AppSearch service
generated code size by 2.2MB!

Background: A profile is a list of classes and methods that are
frequently used, which ART uses to compile parts of the code ahead of
time, and the rest of the code can be interpreted and Jitted on the fly
if needed, this improves app-startup, memory and disk space
significantly!

For system server modules (like AppSearch), the current behavoir is to
Speed compile (compile everything), which for sure compiles unnecessary
code. By moving the speed-profile mode we are saving around 2.2MB of
disk space which also decreases the load on memory!

The new behavior is guarded by a new build feature
`RELEASE_ART_COMPILE_APP_SEARCH_SERVICE_SPEED_PROFILE`

Bug: 241823638
Flag: build.RELEASE_ART_COMPILE_APP_SEARCH_SERVICE_SPEED_PROFILE
Test: `m`
Test: with flag on: go/art-benchmark?p=BootImageProfileId:54918&f=CandidateName:in:54920 (code size reduction by -2.2MB)
Test: with flag off: go/art-benchmark?p=BootImageProfileId:54918&f=CandidateName:in:54919 (no code change or metric changes)
Change-Id: I09075df2ac7025342071c5da33a719facda201fe