Dump failed migration status in dumpsys
It also includes how many times we have run data migration. This way,
 we can decide if we want to use that to make decions if we wanna
try failed puts.

Also change the way how we decide if migrations needs to be run. With
this change, we only check icing/version file exists no matter if data
migration status file is present or not. It is a safer change as we do
wipe the icing directory if migration runs successfully.

With this change, even if we have rollback and rollfoward, we can still
make migration run.

Bug: 430610163
Test: atest DataMigrationUtilTest
Flag: com.android.appsearch.flags.enable_isolated_storage
Change-Id: I7c5b63e4f1031193d5eae81411a67a31a8979b7f