Always create the migration status file on VM initialization.
This change ensures that a migration status file is always present after the isolated storage service is initialized, even if data migration doesn't run. This file can be used as a marker to identify and remove the VM if it is later disabled. The utility function for writing the migration status has been renamed and updated to handle empty status files gracefully.

Bug: 430643956
Test: check isolated storage vm folder is deleted after disablement
Test: atest AppSearchSessionCtsTest
Test: atest --rebuild-module-info CtsAppSearchTestCases AppSearchServicesTests
      AppSearchMockingServicesTests AppSearchCoreTests
Flag: com.android.appsearch.flags.enable_isolated_storage
Change-Id: I781f55797749ba88b75adb64e8ac88427c6c1543