Fix bad pattern in DataMigrationUtilTest.
DataMigrationUtilTest#verifyVMResetBeforeMigration uses
swapIcingSearchEngineLocked without ensuring that the two search engines
are in an equivalent state. The DataMigrationUtil (the only production
code that swaps) *does* guarantee this.

Since the test wants to switch between two Icing instances that are NOT
in the same state (as a test convenience for putting data into Icing),
we will recreate AppSearchImpl after the swap to ensure that all of
AppSearchImpl's cached data aligns with the Icing instance that it is
using.

Bug: 436813583
Flag: com.android.appsearch.flags.enable_early_set_schema_exit
Test: atest AppSearchServicesTests:com.android.server.appsearch.isolated_storage_service.DataMigrationUtilTest#verifyVMResetBeforeMigration

Change-Id: I2ac0ce8b20227f0994f35213b3fdb969cff373c0