Log when isolated storage data migration is skipped on the first VM run.
This change adds logging to track instances where the isolated storage VM is launched for the first time, but data migration is not required.

From our dashboard we only see a small number of log entries for data migration. It might be due to migration is skipped(e.g. if device gets wiped during upgrade). Adding this log entry will help us understand if that's the case, or there is something wrong on how we log the stats.

Bug: 430289015
Test: atest AppSearchSessionCtsTest
Flag: com.android.appsearch.flags.enable_isolated_storage
Change-Id: I17cd18e09ca43145a0a234f681bdd6f309b58151