Fixes flaky tests
What we want the test to do was pretend to already schedule a job info, so that when we run the sync, it would detect that the correct job is already scheduled, and so skip scheduling, which we check with `verify(mockJobScheduler, never()).schedule(any());`

However, we weren't correctly mocking the JobInfo. So when we would check if the correct job was already scheduled, equals would always return false, and so it would schedule the job.

```
if (jobInfo.equals(pendingJobInfo)) { // This was always false
    return;
}
jobScheduler.schedule(jobInfo);
```

On top of that, we didn't have the latch set up correctly. We would setup the latch, call doUpdate (synchronous method), which would trigger the update and release the latch, THEN call updateAsync and wait for the latch. Because the synchronous method already modified the documents and released the latch, we would then immediately call verify().

95% of the time, we would reach the verify() line BEFORE schedule was called, causing the test to pass. But sometimes we would reach it AFTER schedule was called, failing the test. Setting firstRun=true caused an early return because of the first doUpdate call in the test (skip if ran too recently). This would increase the chance that we asynchronously call schedule() before the verify() line, causing the test to fail 5% of the time.

With firstRun=false, it's less likely as the sync call has more latency because we'll be calling PackageManager and AppSearch to see what we need to update. So schedule was more likely to be called after the verify line.

Also firstRun=true is not the right setup here because we want to index the test documents, and it will skip indexing the test documents if it happens too quickly after the first doUpdate() call.

To fix this, we can mock the proper JobInfo, remove the doUpdate call, and set firstRun=false.

Bug: 439970103
Flag: EXEMPT TEST_ONLY
Test: AppsIndexerUserInstanceTest
Change-Id: Icd2838230bc7b9d00f92e1249d5fe6b28cf5d50f