Renable AppSearchUserInstanceManagerTest by conditionally running tests.
AppSearchUserInstanceManagerTest was previously disabled in order to
unblock partner OEM launch. The cancel user instance tests were failing
specifically on HSUM builds because they assumed that the user handle
is 0, which not true for HSUM builds.

To resolve we added an assumeTrue so that the tests were only true for
devices where VM is enabled i.e. devices that are not HSUM.

Bug: 438037081
Flag: EXEMPT TEST_ONLY
Test: atest AppSearchUserInstanceManagerTest
Change-Id: I308b2e77031274a20dbca1676c3fbe65811fdc73