Change Commit and Remove blob API write to parcel model.
We are using WRITE_PARCEL_MODE_MARSHALL_WRITE_IN_BLOB in Commit and
Remove blob response object, and wrap the output in a
AppSearchResultParcelV2. The outcome of writeBlob will contains a fd
to Ashmem. We should only use WRITE_PARCEL_MODE_MARSHALL_WRITE_IN_BLOB
in the AppSearchResultParcelV2 which is the object we are sending via
binder, not anything inside of it.

Bug: 444290368
UpstreamCL: 3765237
Test: AppSearchBlobCtsTest
Flag: com.android.appsearch.flags.enable_directly_write_commit_remove_blob_response
Change-Id: Ia2647505b7abbb2b9d71bb25e3ccc02a25243955