Read ro.hw_timeout_multiplier for scaling timeout
Instead of reading nonprotected_appsearch to check if we are running on
CF, read ro.hw_timeout_multiplier which is set based on device.
ro.hw_timeout_multiplier is a no-op on devices that don't set it, and is
usually used by emulators to account for slowness.

Bug: N/A
Test: TH
Flag: NONE refactor
Change-Id: I7fdbc40b621c210c1a407336b7a18e73a8b84977